+++ command/connect/envoy/envoy_test.go
@@ -84,6 +84,7 @@ func TestGenerateConfig(t *testing.T) {
 				ProxyID:               "test-proxy",
 				AgentAddress:          "127.0.0.1",
 				AgentPort:             "8502", // Note this is the gRPC port
+				AdminAccessLogPath:    "/dev/null",
 				AdminBindAddress:      "127.0.0.1",
 				AdminBindPort:         "19000",
 				LocalAgentClusterName: xds.LocalAgentClusterName,
@@ -99,6 +100,7 @@ func TestGenerateConfig(t *testing.T) {
 				ProxyID:               "test-proxy",
 				AgentAddress:          "127.0.0.1",
 				AgentPort:             "8502", // Note this is the gRPC port
+				AdminAccessLogPath:    "/dev/null",
 				AdminBindAddress:      "127.0.0.1",
 				AdminBindPort:         "19000",
 				LocalAgentClusterName: xds.LocalAgentClusterName,
@@ -116,6 +118,7 @@ func TestGenerateConfig(t *testing.T) {
 				ProxyID:               "test-proxy",
 				AgentAddress:          "127.0.0.1",
 				AgentPort:             "8502", // Note this is the gRPC port
+				AdminAccessLogPath:    "/dev/null",
 				AdminBindAddress:      "127.0.0.1",
 				AdminBindPort:         "19000",
 				LocalAgentClusterName: xds.LocalAgentClusterName,
@@ -136,6 +139,7 @@ func TestGenerateConfig(t *testing.T) {
 				ProxyID:               "test-proxy",
 				AgentAddress:          "127.0.0.1",
 				AgentPort:             "8502", // Note this is the gRPC port
+				AdminAccessLogPath:    "/dev/null",
 				AdminBindAddress:      "127.0.0.1",
 				AdminBindPort:         "19000",
 				LocalAgentClusterName: xds.LocalAgentClusterName,
@@ -156,6 +160,7 @@ func TestGenerateConfig(t *testing.T) {
 				ProxyID:               "test-proxy",
 				AgentAddress:          "127.0.0.1",
 				AgentPort:             "8502", // Note this is the gRPC port
+				AdminAccessLogPath:    "/dev/null",
 				AdminBindAddress:      "127.0.0.1",
 				AdminBindPort:         "19000",
 				LocalAgentClusterName: xds.LocalAgentClusterName,
@@ -175,6 +180,7 @@ func TestGenerateConfig(t *testing.T) {
 				// to do.
 				AgentAddress:          "127.0.0.1",
 				AgentPort:             "9999",
+				AdminAccessLogPath:    "/dev/null",
 				AdminBindAddress:      "127.0.0.1",
 				AdminBindPort:         "19000",
 				LocalAgentClusterName: xds.LocalAgentClusterName,
@@ -194,6 +200,25 @@ func TestGenerateConfig(t *testing.T) {
 				// to do.
 				AgentAddress:          "127.0.0.1",
 				AgentPort:             "9999",
+				AdminAccessLogPath:    "/dev/null",
+				AdminBindAddress:      "127.0.0.1",
+				AdminBindPort:         "19000",
+				LocalAgentClusterName: xds.LocalAgentClusterName,
+			},
+		},
+		{
+			Name:  "access-log-path",
+			Flags: []string{"-proxy-id", "test-proxy", "-admin-access-log-path", "/some/path/access.log"},
+			Env:   []string{},
+			WantArgs: BootstrapTplArgs{
+				ProxyCluster: "test-proxy",
+				ProxyID:      "test-proxy",
+				// Should resolve IP, note this might not resolve the same way
+				// everywhere which might make this test brittle but not sure what else
+				// to do.
+				AgentAddress:          "127.0.0.1",
+				AgentPort:             "8502",
+				AdminAccessLogPath:    "/some/path/access.log",
 				AdminBindAddress:      "127.0.0.1",
 				AdminBindPort:         "19000",
 				LocalAgentClusterName: xds.LocalAgentClusterName,
@@ -230,6 +255,7 @@ func TestGenerateConfig(t *testing.T) {
 				ProxyID:               "test-proxy",
 				AgentAddress:          "127.0.0.1",
 				AgentPort:             "8502",
+				AdminAccessLogPath:    "/dev/null",
 				AdminBindAddress:      "127.0.0.1",
 				AdminBindPort:         "19000",
 				LocalAgentClusterName: xds.LocalAgentClusterName,
@@ -261,6 +287,7 @@ func TestGenerateConfig(t *testing.T) {
 				ProxyID:               "test-proxy",
 				AgentAddress:          "127.0.0.1",
 				AgentPort:             "8502",
+				AdminAccessLogPath:    "/dev/null",
 				AdminBindAddress:      "127.0.0.1",
 				AdminBindPort:         "19000",
 				LocalAgentClusterName: xds.LocalAgentClusterName,
@@ -297,6 +324,7 @@ func TestGenerateConfig(t *testing.T) {
 				ProxyID:               "test-proxy",
 				AgentAddress:          "127.0.0.1",
 				AgentPort:             "8502",
+				AdminAccessLogPath:    "/dev/null",
 				AdminBindAddress:      "127.0.0.1",
 				AdminBindPort:         "19000",
 				LocalAgentClusterName: xds.LocalAgentClusterName,
@@ -320,6 +348,7 @@ func TestGenerateConfig(t *testing.T) {
 				ProxyID:               "test-proxy",
 				AgentAddress:          "127.0.0.1",
 				AgentPort:             "8502",
+				AdminAccessLogPath:    "/dev/null",
 				AdminBindAddress:      "127.0.0.1",
 				AdminBindPort:         "19000",
 				LocalAgentClusterName: xds.LocalAgentClusterName,
@@ -373,6 +402,7 @@ func TestGenerateConfig(t *testing.T) {
 				ProxyID:               "test-proxy",
 				AgentAddress:          "127.0.0.1",
 				AgentPort:             "8502",
+				AdminAccessLogPath:    "/dev/null",
 				AdminBindAddress:      "127.0.0.1",
 				AdminBindPort:         "19000",
 				LocalAgentClusterName: xds.LocalAgentClusterName,