+++ ui-v2/app/services/repository/proxy.js
@@ -19,7 +19,7 @@ export default RepositoryService.extend({
     }
     return this.get('store').query(this.getModelName(), query);
   },
-  findInstanceBySlug: function(id, slug, dc, configuration) {
+  findInstanceBySlug: function(id, node, slug, dc, configuration) {
     return this.findAllBySlug(slug, dc, configuration).then(function(items) {
       let res = {};
       if (get(items, 'length') > 0) {