+++ ui-v2/tests/unit/controllers/settings-test.js
@@ -2,7 +2,7 @@ import { moduleFor, test } from 'ember-qunit';
 
 moduleFor('controller:settings', 'Unit | Controller | settings', {
   // Specify the other units that are required for this test.
-  needs: ['service:settings', 'service:dom'],
+  needs: ['service:settings', 'service:dom', 'service:timeout'],
 });
 
 // Replace this with your real tests.